+++ src/main/java/com/cxq/community/community/service/QuestionService.java
@@ -69,7 +69,7 @@ public class QuestionService {
         return paginationDTO;
     }
 
-    public PaginationDTO list(Integer userId, Integer page, Integer size) {
+    public PaginationDTO list(Long userId, Integer page, Integer size) {
         PaginationDTO paginationDTO = new PaginationDTO();
 
         Integer totalPage;
@@ -113,7 +113,7 @@ public class QuestionService {
         return paginationDTO;
     }
 
-    public QuestionDTO getById(Integer id) {
+    public QuestionDTO getById(Long id) {
         Question question= questionMapper.selectByPrimaryKey(id);
         if(question==null){
             throw new CustomizeException(CustomizeErrorCode.QUESTION_NOT_FOUND);
@@ -136,7 +136,7 @@ public class QuestionService {
         }
     }
 
-    public void incView(Integer id, HttpServletResponse response, HttpServletRequest request) {
+    public void incView(Long id, HttpServletResponse response, HttpServletRequest request) {
 
         Question question = new Question();
         question.setId(id);
